Rename configureState to configureInitialState
