+++ src/lib/utils/configureInitialState.js
